Move class FakeResponse to test_api
